Simplify the database
Lots of changes here. The database seemed too complicated/large for its
purpose. I've simplified it down to scriptPubKey_Hash and txid columns.
This will require indexing the block chain from scratch, but I think it
will be much faster this time.

I've also started fixing the timeout for requests on addresses with many
transactions by limiting it to the config.maxTransactionsPerAddress
variable.